Make presto-udfs compatible with presto v0.157
New versions of presto significantly limit what functions can be added from outside Presto.
Due to these limitations functions like NVL and HASH which used advanced concepts of code-gen
and variable arity are not supported.
There is also a change in the way udfs need to be written, we can follow the presto's standard
annotation based mechanism to write all methods and the jar crawler is not needed anymore to
create sclar/aggregate/window functions based on the locations of classes in the repo.